+++ plugin-server/src/worker/plugins/loadPluginsFromDB.ts
@@ -1,0 +1,71 @@
+import { PluginAttachment } from '@posthog/plugin-scaffold'
+
+import { Hub, Plugin, PluginConfig, PluginConfigId, PluginId, TeamId } from '../../types'
+import { getPluginAttachmentRows, getPluginConfigRows, getPluginRows } from '../../utils/db/sql'
+
+export async function loadPluginsFromDB(
+    hub: Hub
+): Promise<Pick<Hub, 'plugins' | 'pluginConfigs' | 'pluginConfigsPerTeam'>> {
+    const startTimer = new Date()
+    const pluginRows = await getPluginRows(hub)
+    const plugins = new Map<PluginId, Plugin>()
+
+    for (const row of pluginRows) {
+        plugins.set(row.id, row)
+    }
+    hub.statsd?.timing('load_plugins.plugins', startTimer)
+
+    const pluginAttachmentTimer = new Date()
+    const pluginAttachmentRows = await getPluginAttachmentRows(hub)
+    const attachmentsPerConfig = new Map<TeamId, Record<string, PluginAttachment>>()
+    for (const row of pluginAttachmentRows) {
+        let attachments = attachmentsPerConfig.get(row.plugin_config_id!)
+        if (!attachments) {
+            attachments = {}
+            attachmentsPerConfig.set(row.plugin_config_id!, attachments)
+        }
+        attachments[row.key] = {
+            content_type: row.content_type,
+            file_name: row.file_name,
+            contents: row.contents,
+        }
+    }
+    hub.statsd?.timing('load_plugins.plugin_attachments', pluginAttachmentTimer)
+
+    const pluginConfigTimer = new Date()
+    const pluginConfigRows = await getPluginConfigRows(hub)
+
+    const pluginConfigs = new Map<PluginConfigId, PluginConfig>()
+    const pluginConfigsPerTeam = new Map<TeamId, PluginConfig[]>()
+
+    for (const row of pluginConfigRows) {
+        const plugin = plugins.get(row.plugin_id)
+        if (!plugin) {
+            continue
+        }
+        const pluginConfig: PluginConfig = {
+            ...row,
+            plugin: plugin,
+            attachments: attachmentsPerConfig.get(row.id) || {},
+            vm: null,
+        }
+        pluginConfigs.set(row.id, pluginConfig)
+
+        if (!row.team_id) {
+            console.error(`🔴 PluginConfig(id=${row.id}) without team_id!`)
+            continue
+        }
+
+        let teamConfigs = pluginConfigsPerTeam.get(row.team_id)
+        if (!teamConfigs) {
+            teamConfigs = []
+            pluginConfigsPerTeam.set(row.team_id, teamConfigs)
+        }
+        teamConfigs.push(pluginConfig)
+    }
+    hub.statsd?.timing('load_plugins.plugin_configs', pluginConfigTimer)
+
+    hub.statsd?.timing('load_plugins.total', startTimer)
+
+    return { plugins, pluginConfigs, pluginConfigsPerTeam }
+}
