+++ plugin-server/tests/worker/capabilities.test.ts
@@ -1,24 +1,31 @@
 import { Hub, LogLevel, PluginCapabilities } from '../../src/types'
 import { createHub } from '../../src/utils/db/hub'
+import { loadSchedule } from '../../src/worker/plugins/loadSchedule'
+import { setupPlugins } from '../../src/worker/plugins/setup'
 import { getVMPluginCapabilities, shouldSetupPluginInServer } from '../../src/worker/vm/capabilities'
 import { createPluginConfigVM } from '../../src/worker/vm/vm'
 import { pluginConfig39 } from '../helpers/plugins'
 
+jest.mock('../../src/worker/plugins/loadSchedule')
+jest.mock('../../src/worker/plugins/loadPluginsFromDB', () => ({
+    loadPluginsFromDB: () => Promise.resolve({ plugins: [], pluginConfigs: [], pluginConfigsPerTeam: [] }),
+}))
+
 describe('capabilities', () => {
+    let hub: Hub
+    let closeHub: () => Promise<void>
+
+    beforeAll(async () => {
+        console.info = jest.fn() as any
+        console.warn = jest.fn() as any
+        ;[hub, closeHub] = await createHub({ LOG_LEVEL: LogLevel.Warn })
+    })
+
+    afterAll(async () => {
+        await closeHub()
+    })
+
     describe('getVMPluginCapabilities()', () => {
-        let hub: Hub
-        let closeHub: () => Promise<void>
-
-        beforeAll(async () => {
-            console.info = jest.fn() as any
-            console.warn = jest.fn() as any
-            ;[hub, closeHub] = await createHub({ LOG_LEVEL: LogLevel.Warn })
-        })
-
-        afterAll(async () => {
-            await closeHub()
-        })
-
         function getCapabilities(indexJs: string): PluginCapabilities {
             const vm = createPluginConfigVM(hub, pluginConfig39, indexJs)
             return getVMPluginCapabilities(vm)
@@ -149,6 +156,16 @@ describe('capabilities', () => {
                 const shouldSetupPlugin = shouldSetupPluginInServer({ processAsyncHandlers: true }, { methods: [] })
                 expect(shouldSetupPlugin).toEqual(false)
             })
+        })
+    })
+
+    describe('setupPlugins()', () => {
+        it('calls loadSchedule only if pluginScheduledTasks is true', async () => {
+            await setupPlugins({ ...hub, capabilities: { pluginScheduledTasks: false } })
+            expect(loadSchedule).not.toHaveBeenCalled()
+
+            await setupPlugins({ ...hub, capabilities: { pluginScheduledTasks: true } })
+            expect(loadSchedule).toHaveBeenCalled()
         })
     })
 })
